+++ CHANGELOG.md
@@ -17,6 +17,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Prevent throwing in `lighthouse:ide-helper` when no custom directives are defined https://github.com/nuwave/lighthouse/pull/948
 
+## Changed
+
+- Validate requirements for argument definitions of `@delete`, `@forceDelete` and `@restore`
+  during schema build time https://github.com/nuwave/lighthouse/pull/941
+
 ## [4.2.1](https://github.com/nuwave/lighthouse/compare/v4.2.0...v4.2.1)
 
 ### Fixed